Added LEIN name coding
I added the LEIN name coding procedure.  This was added mostly because
it was easy to implement.  I believe, though I don't know for certain,
that LEIN is a reference to the Michigan Law Enforcement Information
Network.  But the only documentation for this I can find is cited in the
R help page.